pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
pydantic-settings = "^2.2.1"
cryptography = "^42.0.7"
alembic = "^1.13.1"
aiomysql = "^0.2.0"
black = "^24.4.2"
mypy = "^1.10.0"
sqladmin = {extras = ["full"], version = "^0.17.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
